+++ packages/map-template/src/components/Sidebar/Modal/Modal.jsx
@@ -33,8 +33,9 @@ function Modal({ children, isOpen }) {
      * If the height of the content is bigger than the height of the modal, the fullHeight should be set.
      */
     useEffect(() => {
-        if (!contentRef) return;
+        if (!contentRef.current) return;
         const observer = new MutationObserver(() => {
+            if (!contentRef.current) return;
             const contentHeight = contentRef.current.clientHeight;
             const modalHeight = modalRef.current?.clientHeight;
             setFullHeight(contentHeight > modalHeight);