+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.99.21] - 2026-08-28
+
+### Changed
+
+- Updated the MapsIndoors Web SDK to version 4.59.4.
+
 ## [1.99.20] - 2026-08-26
 
 ### Changed